# api/openapi.yaml
openapi: 3.1.0
info:
  title: cyber-databrew API
  version: 0.1.0
servers:
  - url: http://localhost:8080
security:
  - GraceToken: []
components:
  securitySchemes:
    GraceToken:
      type: apiKey
      in: header
      name: X-Grace-Token
    AdminToken:
      type: apiKey
      in: header
      name: X-Admin-Token
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code: { type: string }
        message: { type: string }
        request_id: { type: string }
        details:
          type: object
          additionalProperties: true
    AssetID:
      type: string
      minLength: 8
      maxLength: 8
      pattern: '^[0-9A-Za-z]{8}$'
      description: Primary key for assets (8 alphanumeric ASCII characters).
    McapFileID:
      type: string
      minLength: 8
      maxLength: 8
      pattern: '^[0-9A-Za-z]{8}$'
      description: Primary key for mcap_files (same format as asset_id).
...
    AssetMcapLocator:
      type: object
      description: >
        Joined view of an asset's time window and its underlying MCAP object,
        for preview/streaming services that drive byte-range reads.
      required: [asset_id, lifecycle_state, mcap, window, version, updated_at]
      properties:
        asset_id: { $ref: '#/components/schemas/AssetID' }
        lifecycle_state: { type: string }
        mcap:
          type: object
          required: [mcap_file_id, mcap_uri, size_bytes]
          properties:
            mcap_file_id: { $ref: '#/components/schemas/McapFileID' }
            mcap_uri:
              type: string
              description: "Storage URI (e.g. gs://bucket/path/to.mcap). Caller signs/reads as needed."
            size_bytes: { type: integer, format: int64 }
            raw_hash_md5: { type: string, nullable: true }
        window:
          type: object
          required: [start_timestamp_ns, end_timestamp_ns, duration_ms]
          properties:
            start_timestamp_ns: { type: integer, format: int64 }
            end_timestamp_ns:   { type: integer, format: int64 }
            duration_ms:        { type: integer, format: int64 }
        version: { type: integer, format: int64 }
        updated_at: { type: string, format: date-time }
...
paths:
...
  /api/v1/assets/{id}/mcap-locator:
    get:
      summary: Get asset MCAP locator (time window + MCAP object)
      description: |
        Returns the underlying MCAP object reference and the asset's time window
        in a single payload, intended for preview/streaming services that need
        to drive HTTP byte-range reads against the original MCAP without
        signing URLs through this API.

        The endpoint is read-only. URL signing, byte-range planning and video
        muxing are responsibilities of downstream services (e.g.
        mcap-preview-service).
      parameters:
        - in: path
          name: id
          required: true
          schema: { $ref: '#/components/schemas/AssetID' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AssetMcapLocator' }
        '404':
          description: asset or mcap file not found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '409':
          description: |
            asset lifecycle_state is not preview-enabled for this endpoint
            (allowed created, ready, delivered, archived, superseded),
            or asset has no linked mcap file
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '503':
          description: mcap repository not configured on this server
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
...